+++ src/Compute/Compute/VirtualMachine/Config/SetAzureVMOperatingSystemCommand.cs
@@ -249,9 +249,26 @@ namespace Microsoft.Azure.Commands.Compute
             ParameterSetName = LinuxParamSet,
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "Choose one of the following settings: 'Manual', 'AutomaticByOS', or 'AutomaticByPlatform'")]
-        [ValidateNotNullOrEmpty]
         [PSArgumentCompleter("Manual", "AutomaticByOS", "AutomaticByPlatform")]
         public string PatchMode { get; set; }
+
+        [Parameter(
+            ParameterSetName = WindowsParamSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Enables customers to patch their Azure VMs without requiring a reboot. For enableHotpatching, the 'provisionVMAgent' must be set to true and 'patchMode' must be set to 'AutomaticByPlatform'.")]
+        [Parameter(
+            ParameterSetName = WinRmHttpsParamSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Enables customers to patch their Azure VMs without requiring a reboot. For enableHotpatching, the 'provisionVMAgent' must be set to true and 'patchMode' must be set to 'AutomaticByPlatform'.")]
+        [Parameter(
+            ParameterSetName = WindowsDisableVMAgentParamSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Enables customers to patch their Azure VMs without requiring a reboot. For enableHotpatching, the 'provisionVMAgent' must be set to true and 'patchMode' must be set to 'AutomaticByPlatform'.")]
+        [Parameter(
+            ParameterSetName = WindowsDisableVMAgentWinRmHttpsParamSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Enables customers to patch their Azure VMs without requiring a reboot. For enableHotpatching, the 'provisionVMAgent' must be set to true and 'patchMode' must be set to 'AutomaticByPlatform'.")]
+        public SwitchParameter EnableHotpatching { get; set; }
 
         // Linux Parameter Sets
         [Parameter(
@@ -365,7 +382,14 @@ namespace Microsoft.Azure.Commands.Compute
                         this.VM.OSProfile.WindowsConfiguration.PatchSettings = new PatchSettings();
                     }
                     this.VM.OSProfile.WindowsConfiguration.PatchSettings.PatchMode = this.PatchMode;
+
+                    if (this.IsParameterBound(c => c.EnableHotpatching))
+                    {
+                        this.VM.OSProfile.WindowsConfiguration.PatchSettings.EnableHotpatching = this.EnableHotpatching;
+                    }
                 }
+
+
             }
 
             WriteObject(this.VM);